opm-ci-midstream-recombination-tests: Add emails, separate cleanup
This commit separates the cleanup of hosts into it's own builder so that
we can correctly monitor the status of it and issue emails. This also
allows us to set options specifically for the main build run. The first
of which is to ignore the exit codes of the main builder.

This commit also adds emails if the jobs to not run as expected. Since
build failure meaning is unclear at this point, alerting is the best we
can do.

Change-Id: I06a3789f625f19057630926eae01686900300d8b