Move cleanup and vote-recombinations inside khaleesi runs loop
Number of recombinations to test may be high, and executing cleanup and
vote at the end of the build script may result in resource locking for
too much time, and problems in the loop may prevent all the
recombinations to be voted and all the provisioned node to be held. Move
cleanup and vote-recombinations are run at the end of every loop, not
only at the end of the build script.

Change-Id: I9a79d471754f2881243be3cd5a6c0df69dc6ac83